PR 150: Increase the ancient level sql data type
 - Increase the ancient level sql data type

+++ ClickerHeroesTrackerWebsite/ClickerHeroesTrackerWebsite/Services/Database/Schemas/SqlServer/AncientLevels.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TABLE [dbo].[AncientLevels] (
-    [AncientId] TINYINT NOT NULL,
-    [UploadId]  INT     NOT NULL,
-    [Level]     REAL    NOT NULL,
+    [AncientId] TINYINT    NOT NULL,
+    [UploadId]  INT        NOT NULL,
+    [Level]     FLOAT (53) NOT NULL,
     CONSTRAINT [FK_AncientLevels_ToUploads] FOREIGN KEY ([UploadId]) REFERENCES [dbo].[Uploads] ([Id])
 );
